requirements/dev.txt
celery<4
Django<1.12
django-environ<0.5
djangorestframework<4
ipython==5.5.0
ipython-genutils==0.2.0
factory_boy